% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_consistency.R
\name{calculate_consistency}
\alias{calculate_consistency}
\title{Calculate Consistency of Relative vs Absolute Changes}
\usage{
calculate_consistency(fold_changes = NULL, relative_fc = NULL,
  log_scale = FALSE)
}
\arguments{
\item{fold_changes}{a vector of absolute fold changes}

\item{relative_fc}{a vector of perceived relative fold changes}

\item{log_scale}{if \code{TRUE} (default is \code{FALSE}),
the fold_changes are on the log scale (i.e. 0 is no change);
otherwise the fold_changes are on the absolute scale (i.e. 1
is no change).}
}
\value{
a factor vector listing the consistency for each target.
  the possible levels are "zeroCNs" (zero copy numbers),
  "consistent_noChange", "consistent_change", "zero2change"
  (an absolute 'no change' is perceived as a relative change),
  "change2zero" (an absolute change is perceived as not changing),
  "down2up" (an absolute decrease is perceived as a relative increase),
  or "up2down" (an absolute increase is perceived as a relative decrease)
}
\description{
If the assumption holds true that the overall
composition of (and the total amount of) RNA in each 
sample is the same, then the absolute and relative fold
changes should agree. If the assumption is violated,
then they will not agree. This function compares the 
absolute fold change and the perceived fold change
of the proportions of each target in a collection. It
reports a summary of how many targets were consistent
and how many were inconsistent. It also summarizes the
kind of consistency or inconsistency.
}
