% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_consistency.R
\name{calculate_rel_consistency}
\alias{calculate_rel_consistency}
\title{Calculate Consistency of Logratio Data}
\usage{
calculate_rel_consistency(abs_cns, rel_tpms, denom = NULL)
}
\arguments{
\item{abs_cns, }{an N x 2 matrix or data frame with expected
absolute copy numbers in each of two conditions for N targets}

\item{rel_tpms, }{an N x 2 matrix or data frame with expected
relative TPMs of N targets in each of two conditions}

\item{denom, }{the name(s) of the target(s) to be used as the
denominator for the additive logratio transformation}
}
\value{
a list with three members:
  \itemize{
    \item alr_means: the ALR transformed values of the copy numbers
    \item alr_diff: the difference of logratios between the two conditions
      of the ALR transformed copy numbers
    \item alr_consistency: the result from \code{calculate_consistency}
  }
}
\description{
This function uses \link{calculate_consistency} to
determine if there is consistency in the perceived
fold changes after logratio transformation of the
absolute copy numbers and the relative TPMs. In
theory, all targets should show consistent behavior.
}
