% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sizes.R
\name{calculate_sizes}
\alias{calculate_sizes}
\title{This function calculates size factors for use with the polyester simulation
using real data.
In a negative binomial distribution, if the size is r, then the variance of
the NB random variable is V = mean + mean^2 / r
It is estimated using DESeq2's \code{estimateDispersions} function.
Note: Dispersion = 1 / r}
\usage{
calculate_sizes(counts = NULL, s2c, reads_pt = NULL, fc = NULL,
  single_value = TRUE, min_dispersion = 1e-06)
}
\description{
This function calculates size factors for use with the polyester simulation
using real data.
In a negative binomial distribution, if the size is r, then the variance of
the NB random variable is V = mean + mean^2 / r
It is estimated using DESeq2's \code{estimateDispersions} function.
Note: Dispersion = 1 / r
}
