% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERCC92_seqs}
\alias{ERCC92_seqs}
\title{ERCC ExFold RNA Spike-in Sequences}
\format{A \code{DNAStringSet} object (from Biostrings package) with 92 entries
\describe{
  \item{width}{the length of the sequence; accessed by
    Biostrings::width(ERCC92_seqs)}
  \item{seq}{the sequence; accessed directly by \code{as.character(ERCC92_seqs)}}
  \item{names}{the name of the sequence. The original names used hyphens,
    which are replaced with underscores ('_') here. These are the same as
    the rownames for \link{\code{ERCC92_data}}. Accessed by \code{names(ERCC92_seqs)}}
  \item{elementMetadata}{the metadata for the RNAs; accessed by
    \code{S4Vectors::elementMetadata(ERCC92_seq)}. This contains a data frame
    with 92 rows and 4 columns, with the following column names:
    \itemize{
      \item{ERCC_ID} The ID. This is the same as the names above.
      \item{Genbank} The Genbank entry accession number. Go to
        \url{https://www.ncbi.nlm.nih.gov/nuccore} to search Genbank entries.
      \item{X5prime_assay} The accession number for a predesigned TaqMan(R) Assay,
        with the amplicon in the 5' end of the synthetic RNA. Use
        \url{http://www.thermofisher.com/order/genome-database/browse/gene-expression/keyword/}
        to search for assays.
      \item{X3prime_assay} The accession number for a predesigned TaqMan(R) Assay,
        with the amplicon in the 3' end of the synthetic RNA. Use 
        \url{http://www.thermofisher.com/order/genome-database/browse/gene-expression/keyword/}
        to search for assays.
    }}
}}
\source{
The product page for ERCC ExFold RNA Spike-in Mixes is found at
  \url{https://www.thermofisher.com/order/catalog/product/4456739}

The fasta file used to build the DNAStringSet object is found at
  \url{https://assets.thermofisher.com/TFS-Assets/LSG/manuals/ERCC92.zip}

The metadata was downloaded from
  \url{https://assets.thermofisher.com/TFS-Assets/LSG/manuals/cms_095047.txt}
}
\usage{
ERCC92_seqs
}
\description{
This data contains the sequence for each of the 92 synthetic RNAs included
in the ERCC ExFold spike-in mix (the same RNAs are used for Mix 1 and Mix 2).
These spike-in mixes are the most commonly used way to included spike-in
control sequences into an RNA-Seq experiment. The catalog number at
ThermoFisher is 4456739.
}
\keyword{datasets}
