% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_simulation.R
\name{summarize_abs_sim_runs}
\alias{summarize_abs_sim_runs}
\title{Summarize a group of copy number simulations}
\usage{
summarize_abs_sim_runs(abs_sim_runs_list, num_runs, de_probs, dir_probs)
}
\arguments{
\item{abs_sim_runs_list, }{the output from \code{run_abs_simulation}}

\item{num_runs, }{the number of simulations}

\item{de_probs, }{vector of probabilities of differential expression}

\item{dir_probs, }{vector of probabilities of differential expression
being increased, given a transcript that is changing}
}
\value{
a data frame with the following columns:
  \itemize{
    \item run_num: which simulation?
    \item de_prob: probability of differential expression
    \item dir_prob: probability of DE being up
    \item rel*: the number of transcripts with each category of consistency
      between copy numbers and relative TPMs
    \item rel_apparent_fc: the apparent fold change of relative TPMs of
      transcripts that are not changing
    \item adj*: the number of transcripts with each category of consistency
      between copy numbers and DESeq-normalized counts
    \item adj_apparent_fc: the apparent fold change of DESeq-normalized
      counts for transcripts that are not changing
    \item alr: the number of transcripts that have consistent
      changes after ALR transformation of copy numbers vs relative TPMs
  }
}
\description{
This function summarizes for each simulation the consistency of the
fold changes comparing copy numbers to TPMs, and then the consistency
comparing the ALR transformation of the copy numbers versus that of 
the TPMs.
}
