% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpms_to_expected_reads.R
\name{tpms_to_expected_reads}
\alias{tpms_to_expected_reads}
\title{Convert TPMs to Expected Fragments}
\usage{
tpms_to_expected_reads(trans_abund = NULL, eff_lens = NULL,
  mean_library_size = NULL)
}
\arguments{
\item{trans_abund, }{a N x 2 matrix or data frame of transcript
abundances (in TPMs) for each transcript in each of two conditions.}

\item{eff_lens, }{a vector of effective lengths for each transcript.
the transcripts must be in the same order as what is found in
\code{trans_abund}.}

\item{mean_library_size, }{a single value of the expected size of the
sequencing library (e.g. 60 million fragments)}
}
\value{
a list with two entries:
  \itemize{
    \item expected_reads: an N x 2 matrix of the expected reads for each
      transcript in each condition
    \item fold_changes: the perceived fold changes if one were to use the
      expected reads (unnormalized by library size or transcript effective
      length).
  }
}
\description{
This function takes as input a matrix of transcript abundances
(in TPMs) and converts them to expected fragments for each
transcript in each condition. The expected reads is proportional
to the effective length for the transcript and the expected size
of the library.
}
